flask/requirements.txt
Flask==2.3.2
tensorflow==2.18.0
numpy
opencv-python
Pillow
python-dotenv
Flask-Login
Flask-SQLAlchemy
requests
fpdf
gunicorn